Moved ConfigData to its own util.py file to avoid circular dependencies
